backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
pymongo==4.6.1
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
slowapi==0.1.8
reportlab==4.0.4